feature: SSR for some pages / route

+++ src/app/services/storage/app.local.storage.service.ts
@@ -1,5 +1,6 @@
 import { StorageServiceInterface } from './app.storage.service.interface';
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 
 
 @Injectable({
@@ -9,9 +10,12 @@ export abstract class LocalStorageService implements StorageServiceInterface {
 
   protected abstract nameSpace: string;
 
-  private localStorage = localStorage;
+  private localStorage: Storage | null = null;
 
-  constructor() {
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {
+    if (isPlatformBrowser(this.platformId)) {
+      this.localStorage = localStorage;
+    }
   }
 
   getNameSpace(): string {
@@ -19,31 +23,37 @@ export abstract class LocalStorageService implements StorageServiceInterface {
   }
 
   setItem(key: string, data: string) {
-    localStorage.setItem(
+    if (!this.localStorage) return;
+    this.localStorage.setItem(
       this.nameSpace + key,
       data,
     );
   }
 
   getItem(key: string): string {
-    return localStorage.getItem(this.nameSpace + key)
+    if (!this.localStorage) return null as any; // or return ''
+    return this.localStorage.getItem(this.nameSpace + key) || '';
   }
 
   removeItem(key: string): void {
-    return localStorage.removeItem(this.nameSpace + key)
+    if (!this.localStorage) return;
+    return this.localStorage.removeItem(this.nameSpace + key)
   }
 
   getAllItems(): string[] {
-    const items = [];
+    if (!this.localStorage) return [];
+    const items: string[] = [];
     this.getAllKeys().map((localStorageKey) => {
-      items.push(localStorage.getItem(localStorageKey));
+      const item = this.localStorage!.getItem(localStorageKey);
+      if (item) items.push(item);
     });
     return items;
   }
 
   getAllKeys(): string[] {
-    const localStorageKeys = [];
-    Object.keys(localStorage).map((localStorageKey) => {
+    if (!this.localStorage) return [];
+    const localStorageKeys: string[] = [];
+    Object.keys(this.localStorage).map((localStorageKey) => {
       // If not in the correct namespace move on
       if (localStorageKey.startsWith(this.nameSpace)) {
         localStorageKeys.push(localStorageKey.slice(this.nameSpace.length));
@@ -59,6 +69,7 @@ export abstract class LocalStorageService implements StorageServiceInterface {
   }
 
   clearAllStorage() {
+    if (!this.localStorage) return;
     this.localStorage.clear();
   }
 }
